requirements.txt
requests
python-telegram-bot[job-queue]
beautifulsoup4